(WIP)feat: reuse the exist elements
This commit is working in progress.
Reuse the exist elements means do not let react to regenerate or
remove the existing list DOM which generate the first time when
the new data slice is generate. This can imporve the profile of the
visual list cause react do not need to generate or remove any DOM.